Extract flowable functions to `FlowableUtil`
